moved common/models/code to types/code
