+++ apps/studio/electron/main/code/diff/transform.ts
@@ -10,7 +10,7 @@ import { addClassToNode, replaceNodeClasses } from './style';
 import { updateNodeTextContent } from './text';
 import { assertNever } from '/common/helpers';
 import { type CodeAction, CodeActionType } from '/common/models/actions/code';
-import type { CodeDiffRequest } from '/common/models/code';
+import type { CodeDiffRequest } from '@onlook/types/code';
 import type { TemplateNode } from '/common/models/element/templateNode';
 
 export function transformAst(
