+++ apps/studio/electron/main/code/index.ts
@@ -2,7 +2,7 @@ import { dialog, shell } from 'electron';
 import { PersistentStorage } from '../storage';
 import { formatContent, readFile, writeFile } from './files';
 import { IDE, IdeType } from '/common/ide';
-import type { CodeDiff } from '/common/models/code';
+import type { CodeDiff } from '@onlook/types/code';
 import type { TemplateNode } from '/common/models/element/templateNode';
 
 export async function readCodeBlocks(templateNodes: TemplateNode[]): Promise<string[]> {
