+++ apps/studio/electron/main/events/code.ts
@@ -7,7 +7,7 @@ import { getCodeDiffs } from '../code/diff';
 import { readFile } from '../code/files';
 import { getTemplateNodeChild } from '../code/templateNode';
 import { MainChannels } from '@onlook/types/constants';
-import type { CodeDiff, CodeDiffRequest } from '/common/models/code';
+import type { CodeDiff, CodeDiffRequest } from '@onlook/types/code';
 import type { TemplateNode } from '/common/models/element/templateNode';
 
 export function listenForCodeMessages() {
