+++ apps/studio/src/routes/editor/EditPanel/StylesTab/single/TailwindInput.tsx
@@ -4,7 +4,7 @@ import { sendAnalytics } from '@/lib/utils';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useRef, useState } from 'react';
 import { MainChannels } from '@onlook/types/constants';
-import type { CodeDiffRequest } from '/common/models/code';
+import type { CodeDiffRequest } from '@onlook/types/code';
 import type { TemplateNode } from '/common/models/element/templateNode';
 import { Icons } from '@onlook/ui/icons';
